+++ src/main/java/com/global/api/entities/enums/MobilePaymentMethodType.java
@@ -3,7 +3,8 @@ package com.global.api.entities.enums;
 public enum MobilePaymentMethodType {
 
 	APPLEPAY("apple-pay"),
-	GOOGLEPAY("pay-with-google");
+	GOOGLEPAY("pay-with-google"),
+	CLICK_TO_PAY("click-to-pay");
 		
 	private String value;
 	MobilePaymentMethodType(String value){