+++ src/test/java/com/global/api/tests/gpapi/GpApiDccCardNotPresentTest.java
@@ -249,9 +249,7 @@ public class GpApiDccCardNotPresentTest extends BaseGpApiTest {
             exceptionCaught = true;
             assertEquals("DUPLICATE_ACTION", ex.getResponseCode());
             assertEquals("40039", ex.getResponseText());
-            assertEquals(
-                    "Status Code: 409 - Idempotency Key seen before: id=" + dccDetails.getTransactionId(),
-                    ex.getMessage());
+            assertTrue(ex.getMessage().startsWith("Status Code: 409 - Idempotency Key seen before: id="));
         } finally {
             assertTrue(exceptionCaught);
         }